backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.6.4
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.0
fpdf==1.7.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
gotrue==2.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
isort==5.13.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.31
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.14.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
postgrest==0.16.2
prompt-toolkit==3.0.43
protobuf==4.25.3
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pylint==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
realtime==1.0.2
redis==5.0.3
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.35.1
storage3==0.7.4
streamlit==1.32.2
StrEnum==0.4.15
supabase==2.4.1
supafunc==0.3.3
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.0
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
websockets==11.0.3
yarl==1.9.4
